+++ mtcp/src/mtp_instr.c
@@ -116,7 +116,7 @@ mtp_bp* GetFreeBP(struct tcp_stream *cur_stream){
 tcp_stream* CreateHomaCtx(mtcp_manager_t mtcp, uint32_t cur_ts, uint32_t rpc_ind,
 							uint32_t local_ip, uint16_t local_port, 
 							uint32_t remote_ip, uint16_t remote_port,
-                            uint64_t rpc_id, uint32_t init_seq, uint32_t last_seq, 
+                            uint32_t rpc_id, uint32_t init_seq, uint32_t last_seq, 
                             uint8_t state, uint32_t msg_len, uint32_t cur_offset,
                             uint32_t granted, uint32_t birth,
 							bool rpc_is_client, uint16_t expected_segment_cnt,
